requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.109.0
gevent==23.9.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
pydantic==2.6.0
pydantic_core==2.16.1
sniffio==1.3.0
starlette==0.35.1
tk==0.1.0
typing_extensions==4.9.0
uvicorn==0.27.0.post1
websocket-client==1.7.0
websockets==12.0
zope.event==5.0
zope.interface==6.1
